fix / 좋아요 java 코드 수정

+++ src/main/java/account/process/MembLikeUpd.java
@@ -76,7 +76,7 @@ public class MembLikeUpd {
             System.out.println("업데이트된 행 수: " + updatedRows);
 
             if (updatedRows > 0) {
-                likeCnt = (Integer) session.selectOne("uni-account-mapping.selectOpenLikeCnt", param);
+                likeCnt = session.selectOne("uni-account-mapping.selectOpenLikeCnt", param);
                 System.out.println("LIKE_CNT 값: " + likeCnt);
             }
